requirements.txt
pylint

python-xmp-toolkit
rdflib
exifread
googlemaps
geopy
pyyaml
Pillow-SIMD
mezmorize
